Add files via upload
During the process of elevating my position as a well rounded software engineer it came to my attention that by attending JavaScript meetups that part of the commitment to the group was my participation with presentations.  

These groups attracted all levels of software engineering expertise.  Therefore, my contribution would be welcomed by the beginners as well as the seasoned engineers. Both groups would have suggestions, comments and feedback as to the skill set of the code as well as the selection of the project presented.

Most of the presentations were abstracts from personal projects of interest and not part of any propriety company projects.

I have included the presentation material of these Meetup presentations as they offer insight into the process of communication to the group in attendance.

All of these presentations were hand coded with JavaScript, TypeScript, HTML , JSON and CSS and then VUE.



Thanks,

Vince Schlezes